# fix: TypeScript performance of `DismissibleBanner.tsx`