## Description 📝

- Cleans up  `DismissibleBanner.tsx`'s TypeScript types in hopes to improve `tsc` performance 🧼 
- This isn't very scientific, but I believe it is helping 🧪 

## Preview 📷
![Screenshot 2024-10-09 at 12 20 59 PM](https://github.com/user-attachments/assets/2aa5eb2e-a62c-492b-8bf4-65d10e410683)


| Before  | After   |
| ------- | ------- |
| ![Screenshot 2024-10-09 at 12 18 25 PM](https://github.com/user-attachments/assets/0c4845b3-82a2-4b0b-b528-d4dd705d4e2f)  | ![Screenshot 2024-10-09 at 12 40 13 PM](https://github.com/user-attachments/assets/993fcff6-077d-40a8-9f0f-e605c245ab13) |
|  takes 16 seconds to typecheck | takes 10ms   | 


## How to test 🧪

- Verify type checking passes
- Verify there are no regressions with the component, it's types, and it's behavior
- Profile this if you want to (`yarn tsc --generateTrace ./trace --incremental false` and inspect with `chrome://tracing`)

## As an Author I have considered 🤔

- [ ] 👀 Doing a self review
- [ ] ❔ Our [contribution guidelines](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md)
- [ ] 🤏 Splitting feature into small PRs
- [ ] ➕ Adding a [changeset](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md#writing-a-changeset)
- [ ] 🧪 Providing/Improving test coverage
- [ ] 🔐 Removing all sensitive information from the code and PR description
- [ ] 🚩 Using a feature flag to protect the release
- [ ] 👣 Providing comprehensive reproduction steps
- [ ] 📑 Providing or updating our documentation
- [ ] 🕛 Scheduling a pair reviewing session
- [ ] 📱 Providing mobile support
- [ ] ♿  Providing accessibility support